backend/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
sqlalchemy
requests
python-dotenv
